# api-gateway/docs/openapi.yml
openapi: 3.0.3
info:
  title: Event Booking API Gateway
  description: |
    A comprehensive API for managing events, bookings, and user authentication.
    This API uses NATS messaging for microservice communication and provides
    endpoints for user registration, event management, and booking operations.
  version: 1.0.0
  contact:
    name: API Support
    email: support@eventbooking.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.eventbooking.com
    description: Production server
paths:
...
  /api/bookings:
    get:
      tags:
        - Bookings
      summary: Get all bookings (Admin only)
      description: Retrieves all bookings with optional filtering
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: query
          schema:
            type: string
          description: Filter by event ID
        - name: status
          in: query
          schema:
            type: string
            enum: [confirmed, cancelled, pending]
          description: Filter by booking status
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: Bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Bookings
      summary: Create a new booking
      description: Creates a new booking for an event
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Validation failed or insufficient capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bookings/my-bookings:
    get:
      tags:
        - Bookings
      summary: Get current user's bookings
      description: Retrieves all bookings for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: query
          schema:
            type: string
          description: Filter by event ID
        - name: status
          in: query
          schema:
            type: string
            enum: [confirmed, cancelled, pending]
          description: Filter by booking status
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: User bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bookings/{bookingId}:
    get:
      tags:
        - Bookings
      summary: Get booking by ID
      description: Retrieves detailed information about a specific booking
      security:
        - BearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
          description: The booking ID
      responses:
        '200':
          description: Booking retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Invalid booking ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bookings/{bookingId}/cancel:
    put:
      tags:
        - Bookings
      summary: Cancel a booking
      description: Cancels an existing booking
      security:
        - BearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
          description: The booking ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelBookingRequest'
      responses:
        '200':
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Failed to cancel booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bookings/{bookingId}/validate:
    post:
      tags:
        - Bookings
      summary: Validate a booking (Admin only)
      description: Validates a booking for event entry
      security:
        - BearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
          description: The booking ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateBookingRequest'
      responses:
        '200':
          description: Booking validation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bookings/event/{eventId}:
    get:
      tags:
        - Bookings
      summary: Get bookings for an event (Admin only)
      description: Retrieves all bookings for a specific event
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          description: The event ID
        - name: status
          in: query
          schema:
            type: string
            enum: [confirmed, cancelled, pending]
          description: Filter by booking status
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: Event bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '400':
          description: Invalid event ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bookings/{bookingId}/confirm-payment:
    post:
      tags:
        - Bookings
      summary: Confirm booking payment (Admin only)
      description: Confirms payment for a booking
      security:
        - BearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
          description: The booking ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmPaymentRequest'
      responses:
        '200':
          description: Payment confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Invalid payment information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bookings/admin/stats:
    get:
      tags:
        - Bookings
      summary: Get booking statistics (Admin only)
      description: Retrieves comprehensive booking statistics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Booking statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingStatsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "An error occurred"
        data:
          type: array
          items:
            type: object
          description: Additional error details (validation errors, etc.)

    # Authentication Schemas
...
    CreateBookingRequest:
      type: object
      required:
        - eventId
        - ticketQuantity
        - totalAmount
      properties:
        eventId:
          type: string
          example: "507f1f77bcf86cd799439012"
        ticketQuantity:
          type: integer
          minimum: 1
          example: 2
        totalAmount:
          type: number
          minimum: 0.01
          example: 199.98
        paymentMethod:
          type: string
          example: "credit_card"
    CancelBookingRequest:
      type: object
      properties:
        reason:
          type: string
          example: "Change in plans"
    ValidateBookingRequest:
      type: object
      required:
        - eventId
      properties:
        eventId:
          type: string
          example: "507f1f77bcf86cd799439012"
    ConfirmPaymentRequest:
      type: object
      required:
        - paymentTransactionId
      properties:
        paymentTransactionId:
          type: string
          example: "txn_1234567890"
    Booking:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439013"
        eventId:
          type: string
          example: "507f1f77bcf86cd799439012"
        userId:
          type: string
          example: "507f1f77bcf86cd799439011"
        ticketQuantity:
          type: integer
          example: 2
        totalAmount:
          type: number
          example: 199.98
        status:
          type: string
          enum: [confirmed, cancelled, pending]
          example: "confirmed"
        bookingDate:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        paymentStatus:
          type: string
          enum: [pending, completed, failed, refunded]
          example: "completed"
        paymentMethod:
          type: string
          example: "credit_card"
        event:
          type: object
          properties:
            name:
              type: string
              example: "Tech Conference 2024"
            venue:
              type: string
              example: "Convention Center, Downtown"
            dateTime:
              type: string
              format: date-time
              example: "2024-06-15T09:00:00.000Z"
        user:
          type: object
          properties:
            name:
              type: string
              example: "John Doe"
            email:
              type: string
              format: email
              example: "john@example.com"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    BookingResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Booking'
        message:
          type: string
          example: "Booking retrieved successfully"
    BookingListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            total:
              type: integer
              example: 100
            pages:
              type: integer
              example: 10
        message:
          type: string
          example: "Bookings retrieved successfully"
    BookingStatsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            totalBookings:
              type: integer
              example: 1500
            confirmedBookings:
              type: integer
              example: 1200
            cancelledBookings:
              type: integer
              example: 200
            pendingBookings:
              type: integer
              example: 100
            totalRevenue:
              type: number
              example: 150000.00
            averageBookingValue:
              type: number
              example: 125.50
            topEvents:
              type: array
              items:
                type: object
                properties:
                  eventId:
                    type: string
                    example: "507f1f77bcf86cd799439012"
                  eventName:
                    type: string
                    example: "Tech Conference 2024"
                  bookingCount:
                    type: integer
                    example: 250
                  revenue:
                    type: number
                    example: 24975.00
            monthlyTrends:
              type: array
              items:
                type: object
                properties:
                  month:
                    type: string
                    example: "2024-01"
                  bookings:
                    type: integer
                    example: 120
                  revenue:
                    type: number
                    example: 12000.00
        message:
          type: string
          example: "Booking statistics retrieved successfully"
...